+++ errorprone-checks/src/main/java/org/hyperledger/errorpronechecks/DoNotInvokeMessageDigestDirectly.java
@@ -26,7 +26,6 @@ import com.sun.source.tree.MethodInvocationTree;
 
 @AutoService(BugChecker.class)
 @BugPattern(
-    name = "DoNotInvokeMessageDigestDirectly",
     summary = "Do not invoke MessageDigest.getInstance directly.",
     severity = WARNING,
     linkType = BugPattern.LinkType.NONE)